Worked on many things
