+++ DAO/ICompanyDAO.php
@@ -10,6 +10,6 @@ interface ICompanyDAO
     function GetAll();
     function GetById($number);
     function GetByName($name);
-    function ModifyName($number, $name);
+    function Modify($number, $name, $email, $phone, $address, $cuit, $website, $founded);
     function Delete($number);
 }
