+++ Data/Companies.json
@@ -1,101 +1,79 @@
 [
     {
         "companyId": 1,
-        "name": "Accenture",
-        "email": "jmissart@gmail.com",
-        "phone": "5974846",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "21421525245",
+        "name": "Accenture2",
+        "email": "accenture@gmail.com",
+        "phone": "155123456",
+        "address": "Libertad 1234",
+        "cuit": "20-33938431-2",
         "website": "www.accenture.com",
-        "founded": "2021-10-08",
+        "founded": "2021-10-02",
         "status": true
     },
     {
         "companyId": 2,
         "name": "Globant",
-        "email": "jmissart@gmail.com",
-        "phone": "5974846",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "2142134",
-        "website": "globant.com",
-        "founded": "2021-10-13",
+        "email": "Globant@Globant.com",
+        "phone": "47896542",
+        "address": "Formosa 2525",
+        "cuit": "34-2134214-2",
+        "website": "www.globant.com",
+        "founded": "2021-09-22",
         "status": false
     },
     {
         "companyId": 3,
-        "name": "jaje",
-        "email": "jmissart@gmail.com",
-        "phone": "5974846",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "414123",
-        "website": "www.accenture.com",
-        "founded": "2021-10-21",
+        "name": "Avalith",
+        "email": "Avalith@gmail.com",
+        "phone": "44213245",
+        "address": "Corrientes 5959",
+        "cuit": "23-3214214-6",
+        "website": "www.avalith.com",
+        "founded": "2021-06-19",
         "status": true
     },
     {
         "companyId": 4,
-        "name": "Accenture",
-        "email": "jmissart@gmail.com",
-        "phone": "+54 (223) 528-3129",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "21421",
-        "website": "www.accenture.com",
-        "founded": "2021-10-15",
+        "name": "Infosis",
+        "email": "Infosis@gmail.com",
+        "phone": "43432342",
+        "address": "Av Luro 1999",
+        "cuit": "40-124214-4",
+        "website": "www.infosis.com",
+        "founded": "2021-07-10",
         "status": true
     },
     {
         "companyId": 5,
-        "name": "Julio Missart",
-        "email": "jmissart@gmail.com",
+        "name": "Accenture",
+        "email": "user@myapp.com",
         "phone": "5974846",
         "address": "Av Siempre Viva 1234",
-        "cuit": "2142154",
+        "cuit": "334647247427",
         "website": "www.accenture.com",
-        "founded": "2021-10-20",
+        "founded": "2021-10-02",
         "status": true
     },
     {
         "companyId": 6,
         "name": "Accenture",
         "email": "jmissart@gmail.com",
-        "phone": "142345",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "21421525245",
-        "website": "www.accenture.com",
-        "founded": "2021-10-14",
-        "status": true
-    },
-    {
-        "companyId": 7,
-        "name": "Globant",
-        "email": "ddouthwaite0@goo.gl",
         "phone": "5974846",
         "address": "Av Siempre Viva 1234",
         "cuit": "21421",
         "website": "www.accenture.com",
-        "founded": "2021-10-21",
+        "founded": "2021-10-07",
         "status": true
     },
     {
-        "companyId": 8,
-        "name": "Infosis",
+        "companyId": 7,
+        "name": "Accenture",
         "email": "jmissart@gmail.com",
         "phone": "5974846",
         "address": "Av Siempre Viva 1234",
         "cuit": "21421",
         "website": "www.accenture.com",
-        "founded": "2021-10-11",
-        "status": true
-    },
-    {
-        "companyId": 9,
-        "name": "Accenture",
-        "email": "ddouthwaite0@goo.gl",
-        "phone": "+54 (223) 528-3129",
-        "address": "Av Siempre Viva 1234",
-        "cuit": "21421",
-        "website": "www.accenture.com",
-        "founded": "2021-10-19",
+        "founded": "2021-09-28",
         "status": true
     }
 ]
