--- finalphp2/Config/Config.php
@@ -1,14 +1,0 @@
-<?php namespace Config;
-
-define("ROOT", dirname(__DIR__) . "/");
-//Path to your project's root folder
-define("FRONT_ROOT", "/finalphp2/");
-define("VIEWS_PATH", "Views/");
-define("CSS_PATH", FRONT_ROOT.VIEWS_PATH . "css/");
-define("JS_PATH", FRONT_ROOT.VIEWS_PATH . "js/");
-
-define("DB_HOST", "localhost");
-define("DB_NAME", "joboffers");
-define("DB_USER", "root");
-define("DB_PASS", "root");
-?>
