--- finalphp2/DAO/ICareerDAO.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace DAO;
-
-interface ICareerDAO
-{
-    function GetAll();
-}
