--- finalphp2/DAO/IJobPositionDAO.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace DAO;
-
-interface IJobPositionDAO
-{
-    function GetAll();
-}
