--- finalphp2/DAO/IStudentDAO.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DAO;
-
-interface IStudentDAO
-{
-    function Login($email);
-    function GetAll();
-}
