--- finalphp2/DAO/QueryType.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DAO;
-
-abstract class QueryType
-{
-    const Query = 0;
-    const StoredProcedure = 1;
-}
